Move FileSystemChangeListener to top-level
